python-10/requirements.txt
atomicwrites==1.1.5
attrs==19.1.0
backcall==0.2.0
bleach==3.1.5
certifi==2020.4.5.2
decorator==4.4.2
defusedxml==0.6.0
Django==2.2.13
djangorestframework==3.10.2
entrypoints==0.3
importlib-metadata==1.6.1
ipykernel==5.3.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
MarkupSafe==1.1.1
mistune==0.8.4
more-itertools==4.2.0
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
packaging==20.4
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.6.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
py==1.5.3
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==3.6.2
python-dateutil==2.8.1
pytz==2020.1
pyzmq==18.1.1
qtconsole==4.7.4
QtPy==1.9.0
Send2Trash==1.5.0
six==1.15.0
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
wcwidth==0.2.4
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.1.0
